+++ auth-web/src/components/auth/staff/StaffPendingAccountInvitationsTable.vue
@@ -75,13 +75,13 @@ import { AccessType, Account } from '@/util/constants'
 import { Component, Emit, Mixins, Prop, Vue } from 'vue-property-decorator'
 import { mapActions, mapState } from 'vuex'
 import CommonUtils from '@/util/common-util'
+import { DataOptions } from 'vuetify'
 import { Event } from '@/models/event'
 import { EventBus } from '@/event-bus'
 import { Invitation } from '@/models/Invitation'
 import ModalDialog from '@/components/auth/ModalDialog.vue'
 import { Organization } from '@/models/Organization'
 import PaginationMixin from '@/components/auth/mixins/PaginationMixin.vue'
-import { DataOptions } from 'vuetify'
 
 @Component({
   components: {
